mesa_generator_app/requirements.txt
Flask>=2.0
openai>=1.0
mesa>=1.0
python-dotenv>=0.15 
# Add other dependencies as needed
